requirements.txt
Django==3.1.5
psycopg2==2.8.6
djangorestframework==3.12.2
requests==2.25.1
python-dotenv==0.15.0
django--filter==2.4.0